refactor runner from main function
